More fixes to make 32 bit tests succeed. Make sure `go vet` knows what architecture we're compiling for.

+++ pkg/network/netlink/consumer.go
@@ -103,7 +103,7 @@ func NewConsumer(procRoot string, targetRateLimit int) (*Consumer, error) {
 		pool:            newBufferPool(),
 		workQueue:       make(chan func()),
 		targetRateLimit: targetRateLimit,
-		breaker:         NewCircuitBreaker(targetRateLimit),
+		breaker:         NewCircuitBreaker(int64(targetRateLimit)),
 	}
 	c.initWorker(procRoot)
